Second upload
This is a self-created course plan to achieve the necessary competences to become a data analyst. The courses are from well-known Universities and are all contained in Coursera platform.

The courses have a perfect balance of theory and practice and their structure ensures that the student has mastered the contents after all the sections have been completed.